[Mod] split utility.py into several files
