+++ vnpy/app/algo_trading/engine.py
@@ -5,7 +5,8 @@ from vnpy.trader.event import (
     EVENT_TICK, EVENT_TIMER, EVENT_ORDER, EVENT_TRADE)
 from vnpy.trader.constant import (Direction, Offset, OrderType)
 from vnpy.trader.object import (SubscribeRequest, OrderRequest)
-from vnpy.trader.utility import load_json, save_json, round_to
+from vnpy.utility.math import round_to
+from vnpy.utility.path import load_json, save_json
 
 from .template import AlgoTemplate
