+++ vnpy/app/algo_trading/template.py
@@ -1,7 +1,7 @@
 from vnpy.trader.engine import BaseEngine
 from vnpy.trader.object import TickData, OrderData, TradeData
 from vnpy.trader.constant import OrderType, Offset, Direction
-from vnpy.trader.utility import virtual
+from vnpy.utility.functools import virtual
 
 
 class AlgoTemplate:
