+++ vnpy/app/cta_backtester/engine.py
@@ -8,7 +8,7 @@ from pathlib import Path
 from vnpy.event import Event, EventEngine
 from vnpy.trader.engine import BaseEngine, MainEngine
 from vnpy.trader.constant import Interval
-from vnpy.trader.utility import extract_vt_symbol
+from vnpy.utility.constant import extract_vt_symbol
 from vnpy.trader.object import HistoryRequest
 from vnpy.trader.rqdata import rqdata_client
 from vnpy.trader.database import database_manager
