+++ vnpy/app/cta_strategy/__init__.py
@@ -3,7 +3,8 @@ from pathlib import Path
 from vnpy.trader.app import BaseApp
 from vnpy.trader.constant import Direction
 from vnpy.trader.object import TickData, BarData, TradeData, OrderData
-from vnpy.trader.utility import BarGenerator, ArrayManager
+from vnpy.utility.bar_generator import BarGenerator
+from vnpy.utility.array_manager import ArrayManager
 
 from .base import APP_NAME, StopOrder
 from .engine import CtaEngine
