+++ vnpy/app/cta_strategy/backtesting.py
@@ -17,7 +17,7 @@ from vnpy.trader.constant import (Direction, Offset, Exchange,
                                   Interval, Status)
 from vnpy.trader.database import database_manager
 from vnpy.trader.object import OrderData, TradeData, BarData, TickData
-from vnpy.trader.utility import round_to
+from vnpy.utility.math import round_to
 
 from .base import (
     BacktestingMode,
