+++ vnpy/app/cta_strategy/engine.py
@@ -35,7 +35,9 @@ from vnpy.trader.constant import (
     Offset,
     Status
 )
-from vnpy.trader.utility import load_json, save_json, extract_vt_symbol, round_to
+from vnpy.utility.constant import extract_vt_symbol
+from vnpy.utility.math import round_to
+from vnpy.utility.path import load_json, save_json
 from vnpy.trader.database import database_manager
 from vnpy.trader.rqdata import rqdata_client
