+++ vnpy/app/cta_strategy/strategies/dual_thrust_strategy.py
@@ -6,9 +6,9 @@ from vnpy.app.cta_strategy import (
     BarData,
     TradeData,
     OrderData,
-    BarGenerator,
-    ArrayManager,
 )
+from vnpy.utility.bar_generator import BarGenerator
+from vnpy.utility.array_manager import ArrayManager
 
 
 class DualThrustStrategy(CtaTemplate):
