+++ vnpy/app/cta_strategy/strategies/multi_timeframe_strategy.py
@@ -5,9 +5,9 @@ from vnpy.app.cta_strategy import (
     BarData,
     TradeData,
     OrderData,
-    BarGenerator,
-    ArrayManager,
 )
+from vnpy.utility.bar_generator import BarGenerator
+from vnpy.utility.array_manager import ArrayManager
 
 
 class MultiTimeframeStrategy(CtaTemplate):
