+++ vnpy/app/risk_manager/engine.py
@@ -6,8 +6,7 @@ from vnpy.event import Event, EventEngine, EVENT_TIMER
 from vnpy.trader.engine import BaseEngine, MainEngine
 from vnpy.trader.event import EVENT_TRADE, EVENT_ORDER, EVENT_LOG
 from vnpy.trader.constant import Status
-from vnpy.trader.utility import load_json, save_json
-
+from vnpy.utility.path import load_json, save_json
 
 APP_NAME = "RiskManager"
