+++ vnpy/app/rpc_service/engine.py
@@ -5,7 +5,7 @@ import traceback
 from vnpy.event import Event, EventEngine
 from vnpy.rpc import RpcServer
 from vnpy.trader.engine import BaseEngine, MainEngine
-from vnpy.trader.utility import load_json, save_json
+from vnpy.utility.path import load_json, save_json
 from vnpy.trader.object import LogData
 
 APP_NAME = "RpcService"
