+++ vnpy/gateway/oes/oes_gateway.py
@@ -9,7 +9,7 @@ from threading import Lock, Thread
 from vnpy.trader.gateway import BaseGateway
 from vnpy.trader.object import (CancelRequest, OrderRequest,
                                 SubscribeRequest)
-from vnpy.trader.utility import get_file_path
+from vnpy.utility.path import get_file_path
 from .oes_md import OesMdApi
 from .oes_td import OesTdApi, EXCHANGE_VT2OES
 from .utils import config_template
