+++ vnpy/utility/bar_generator.py
@@ -1,0 +1,153 @@
+from typing import Callable
+
+from vnpy.trader.constant import Interval
+from vnpy.trader.object import TickData, BarData
+
+
+class BarGenerator:
+    """
+    For:
+    1. generating 1 minute bar data from tick data
+    2. generateing x minute bar/x hour bar data from 1 minute data
+
+    Notice:
+    1. for x minute bar, x must be able to divide 60: 2, 3, 5, 6, 10, 15, 20, 30
+    2. for x hour bar, x can be any number
+    """
+
+    def __init__(
+        self,
+        on_bar: Callable,
+        window: int = 0,
+        on_window_bar: Callable = None,
+        interval: Interval = Interval.MINUTE
+    ):
+        """Constructor"""
+        self.bar = None
+        self.on_bar = on_bar
+
+        self.interval = interval
+        self.interval_count = 0
+
+        self.window = window
+        self.window_bar = None
+        self.on_window_bar = on_window_bar
+
+        self.last_tick = None
+        self.last_bar = None
+
+    def update_tick(self, tick: TickData):
+        """
+        Update new tick data into generator.
+        """
+        new_minute = False
+
+        # Filter tick data with 0 last price
+        if not tick.last_price:
+            return
+
+        if not self.bar:
+            new_minute = True
+        elif self.bar.datetime.minute != tick.datetime.minute:
+            self.bar.datetime = self.bar.datetime.replace(
+                second=0, microsecond=0
+            )
+            self.on_bar(self.bar)
+
+            new_minute = True
+
+        if new_minute:
+            self.bar = BarData(
+                symbol=tick.symbol,
+                exchange=tick.exchange,
+                interval=Interval.MINUTE,
+                datetime=tick.datetime,
+                gateway_name=tick.gateway_name,
+                open_price=tick.last_price,
+                high_price=tick.last_price,
+                low_price=tick.last_price,
+                close_price=tick.last_price,
+                open_interest=tick.open_interest
+            )
+        else:
+            self.bar.high_price = max(self.bar.high_price, tick.last_price)
+            self.bar.low_price = min(self.bar.low_price, tick.last_price)
+            self.bar.close_price = tick.last_price
+            self.bar.open_interest = tick.open_interest
+            self.bar.datetime = tick.datetime
+
+        if self.last_tick:
+            volume_change = tick.volume - self.last_tick.volume
+            self.bar.volume += max(volume_change, 0)
+
+        self.last_tick = tick
+
+    def update_bar(self, bar: BarData):
+        """
+        Update 1 minute bar into generator
+        """
+        # If not inited, creaate window bar object
+        if not self.window_bar:
+            # Generate timestamp for bar data
+            if self.interval == Interval.MINUTE:
+                dt = bar.datetime.replace(second=0, microsecond=0)
+            else:
+                dt = bar.datetime.replace(minute=0, second=0, microsecond=0)
+
+            self.window_bar = BarData(
+                symbol=bar.symbol,
+                exchange=bar.exchange,
+                datetime=dt,
+                gateway_name=bar.gateway_name,
+                open_price=bar.open_price,
+                high_price=bar.high_price,
+                low_price=bar.low_price
+            )
+        # Otherwise, update high/low price into window bar
+        else:
+            self.window_bar.high_price = max(
+                self.window_bar.high_price, bar.high_price)
+            self.window_bar.low_price = min(
+                self.window_bar.low_price, bar.low_price)
+
+        # Update close price/volume into window bar
+        self.window_bar.close_price = bar.close_price
+        self.window_bar.volume += int(bar.volume)
+        self.window_bar.open_interest = bar.open_interest
+
+        # Check if window bar completed
+        finished = False
+
+        if self.interval == Interval.MINUTE:
+            # x-minute bar
+            if not (bar.datetime.minute + 1) % self.window:
+                finished = True
+        elif self.interval == Interval.HOUR:
+            if self.last_bar and bar.datetime.hour != self.last_bar.datetime.hour:
+                # 1-hour bar
+                if self.window == 1:
+                    finished = True
+                # x-hour bar
+                else:
+                    self.interval_count += 1
+
+                    if not self.interval_count % self.window:
+                        finished = True
+                        self.interval_count = 0
+
+        if finished:
+            self.on_window_bar(self.window_bar)
+            self.window_bar = None
+
+        # Cache last bar object
+        self.last_bar = bar
+
+    def generate(self):
+        """
+        Generate the bar data and call callback immediately.
+        """
+        self.bar.datetime = self.bar.datetime.replace(
+            second=0, microsecond=0
+        )
+        self.on_bar(self.bar)
+        self.bar = None
