+++ vnpy/utility/functools.py
@@ -1,0 +1,7 @@
+def virtual(func: "callable"):
+    """
+    mark a function as "virtual", which means that this function can be override.
+    any base class should use this or @abstractmethod to decorate all functions
+    that can be (re)implemented by subclasses.
+    """
+    return func
